move initialize_dataframe_geo_index to aggregation helpers
